# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocialPredict API
  version: 0.1.0
  description: >
    HTTP contract for the SocialPredict backend service. This spec is generated
    from the Go handlers and DTOs under backend/server and backend/handlers.
    Routes wrapped by the shared security middleware can return a shared
    transport-level `429` JSON failure envelope response before reaching handler-specific
    JSON envelope logic when the rate limit is exceeded. This middleware
    exception applies across protected routes even when a path omits a repeated
    path-level `429` entry.
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /v0/privateprofile:
    get:
      tags: [Users]
      operationId: getPrivateProfile
      summary: Get private profile
      description: >
        Returns the authenticated user's private profile data. Handler-owned failures use
        the shared JSON failure envelope with INVALID_TOKEN, PASSWORD_CHANGE_REQUIRED,
        USER_NOT_FOUND, METHOD_NOT_ALLOWED, or INTERNAL_ERROR.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Private profile returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUserEnvelopeResponse'
        '401':
          description: Invalid or missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '403':
          description: Password change is required before this route can be used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '405':
          description: HTTP method is not allowed for this route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while fetching profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
  /v0/profilechange/displayname:
    post:
      tags: [Users]
      operationId: updateDisplayName
      summary: Update display name
      description: >
        Protected profile update route migrated to the shared JSON failure envelope.
        Handler-owned failures use INVALID_TOKEN, PASSWORD_CHANGE_REQUIRED,
        INVALID_REQUEST, VALIDATION_FAILED, USER_NOT_FOUND, AUTHORIZATION_DENIED,
        METHOD_NOT_ALLOWED, or INTERNAL_ERROR.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDisplayNameRequest'
      responses:
        '200':
          description: Display name updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUserEnvelopeResponse'
        '400':
          description: Invalid request body or validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '401':
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '403':
          description: Password change is required before this route can be used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '405':
          description: HTTP method is not allowed for this route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while updating display name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
  /v0/profilechange/links:
    post:
      tags: [Users]
      operationId: updatePersonalLinks
      summary: Update personal links
      description: >
        Protected profile update route migrated to the shared JSON failure envelope.
        Handler-owned failures use INVALID_TOKEN, PASSWORD_CHANGE_REQUIRED,
        INVALID_REQUEST, VALIDATION_FAILED, USER_NOT_FOUND, AUTHORIZATION_DENIED,
        METHOD_NOT_ALLOWED, or INTERNAL_ERROR.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePersonalLinksRequest'
      responses:
        '200':
          description: Personal links updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUserEnvelopeResponse'
        '400':
          description: Invalid request body or validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '401':
          description: Authentication failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '403':
          description: Password change is required before this route can be used.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '405':
          description: HTTP method is not allowed for this route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while updating personal links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
  /v0/changepassword:
    post:
      tags: [Users]
      operationId: changePassword
      summary: Change account password
      description: >
        Protected password-change route migrated to the shared JSON failure envelope.
        Handler-owned failures use INVALID_TOKEN, INVALID_REQUEST, VALIDATION_FAILED,
        AUTHORIZATION_DENIED, USER_NOT_FOUND, METHOD_NOT_ALLOWED, or INTERNAL_ERROR.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordEnvelopeResponse'
        '400':
          description: Invalid request body or password requirements not met.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '401':
          description: Invalid or missing token, or the submitted current password does not match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '405':
          description: HTTP method is not allowed for this route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
        '500':
          description: Unexpected server error while changing password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReasonResponse:
      type: object
      description: |
        Shared failure envelope for the scoped W04 contract pass.

        Preferred public reasons for touched envelope endpoints:
        - METHOD_NOT_ALLOWED: a documented handler was called with an unsupported HTTP method.
        - RATE_LIMITED and LOGIN_RATE_LIMITED: shared middleware rejected a request before route-specific handler logic.
        - INVALID_REQUEST: malformed request body, path parameter, or query input.
        - INVALID_TOKEN: missing or invalid bearer token.
        - AUTHORIZATION_DENIED: supplied credentials are not accepted for the requested flow, or the authenticated caller lacks the required privileges.
        - PASSWORD_CHANGE_REQUIRED: request blocked until the caller changes their password.
        - USER_NOT_FOUND, MARKET_NOT_FOUND, and NOT_FOUND: stable not-found outcomes for user, market, and generic content/resource lookups.
        - VALIDATION_FAILED: syntactically valid input that fails route or business validation.
        - MARKET_CLOSED, INSUFFICIENT_BALANCE, NO_POSITION, INSUFFICIENT_SHARES, and DUST_CAP_EXCEEDED: shared market and bet rule failures.
        - INTERNAL_ERROR: unexpected server-side failure.
      properties:
        ok:
          type: boolean
          enum: [false]
        reason:
          type: string
          enum:
            - METHOD_NOT_ALLOWED
            - RATE_LIMITED
            - LOGIN_RATE_LIMITED
            - INVALID_REQUEST
            - INVALID_TOKEN
            - AUTHORIZATION_DENIED
            - PASSWORD_CHANGE_REQUIRED
            - NOT_FOUND
            - USER_NOT_FOUND
            - MARKET_NOT_FOUND
            - VALIDATION_FAILED
            - MARKET_CLOSED
            - INSUFFICIENT_BALANCE
            - NO_POSITION
            - INSUFFICIENT_SHARES
            - DUST_CAP_EXCEEDED
            - INTERNAL_ERROR
          example: VALIDATION_FAILED
      required: [ok, reason]
...
    PrivateUserEnvelopeResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: '#/components/schemas/PrivateUserResponse'
      required: [ok, result]
...
    ChangePasswordEnvelopeResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: '#/components/schemas/ChangePasswordResult'
      required: [ok, result]
    ChangePasswordResult:
      type: object
      properties:
        message:
          type: string
          example: Password changed successfully
      required: [message]
...
    PrivateUserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayname:
          type: string
        usertype:
          type: string
        initialAccountBalance:
          type: integer
          format: int64
        accountBalance:
          type: integer
          format: int64
        personalEmoji:
          type: string
        description:
          type: string
        personalink1:
          type: string
        personalink2:
          type: string
        personalink3:
          type: string
        personalink4:
          type: string
        email:
          type: string
        apiKey:
          type: string
        mustChangePassword:
          type: boolean
...
    ChangeDisplayNameRequest:
      type: object
      properties:
        displayName:
          type: string
...
    ChangePersonalLinksRequest:
      type: object
      properties:
        personalLink1:
          type: string
        personalLink2:
          type: string
        personalLink3:
          type: string
        personalLink4:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...